Create question

+++ src/gui/EditorController.java
@@ -1,18 +1,17 @@
 package gui;
 
-import javafx.collections.FXCollections;
-import javafx.collections.ObservableList;
+
+import javafx.fxml.FXML;
+import javafx.scene.control.*;
 import javafx.event.ActionEvent;
 import javafx.event.EventHandler;
-import javafx.fxml.FXML;
-import javafx.scene.Scene;
-import javafx.scene.control.*;
-import javafx.scene.layout.*;
+
 import quiz.Question;
 import quiz.Topic;
 
 import java.util.ArrayList;
 import java.util.Optional;
+
 
 public class EditorController {
 
@@ -24,7 +23,7 @@ public class EditorController {
     private Button addTopicButton;
 
     private ArrayList<Topic> topics;
-    private Topic selectedTopic;
+    private Topic selectedTopic; // the loaded topic
 
 
     public void initialize() {
@@ -33,6 +32,7 @@ public class EditorController {
     }
 
     private void loadTopicsList() {
+        // remove all items first
         topicsListView.getItems().clear();
 
         // get all topic names and add to listView
@@ -50,7 +50,8 @@ public class EditorController {
             @Override
             public void handle(ActionEvent event) {
                 String name = showInputTextDialog("New Topic", "Enter topic's title");
-                addTopicButton(name);
+                if (name != null)
+                    addTopicButton(name);
             }
         });
     }
@@ -64,8 +65,10 @@ public class EditorController {
         dialog.setContentText("Name:");
 
         Optional<String> result = dialog.showAndWait();
-
-        return result.get();
+        if (result.isPresent())
+            return result.get();
+        else // cancel button
+            return null;
     }
 
     public void loadButton() {
@@ -81,11 +84,12 @@ public class EditorController {
         }
     }
 
-    public void saveButton() {
+    // public void saveButton() {
+    /* not needed anymore
         for (Topic t : topics) {
             t.saveToFile();
         }
-    }
+    }*/
 
     public void closeButton() {
         Main.changeScene("startWindow.fxml");
@@ -98,13 +102,25 @@ public class EditorController {
     }
 
     public void addQuestionButton() {
-        Helper.topUUID = selectedTopic.getId();
+
+        if (selectedTopic == null) {
+            showAlert("Info", "First, Select and Load Topic");
+            return;
+        }
+
+        Helper.topicUUID = selectedTopic.getId();
         Main.changeScene("questionCreation.fxml");
     }
 
     public void deleteTopicButton() {
         // get index of selected object
         int selectedItem = topicsListView.getSelectionModel().getSelectedIndex();
+
+        if (selectedItem == -1) {
+            showAlert("Info", "No Topic selected");
+            return;
+        }
+
         Topic selectedTopic = topics.get(selectedItem);
         selectedTopic.delete();
 
@@ -114,8 +130,13 @@ public class EditorController {
     }
 
     public void deleteQuestionButton() {
-        // get index of selected item
+        // get index of selected item, returns -1 if none is selected
         int selectedItem = questionsListView.getSelectionModel().getSelectedIndex();
+
+        if (selectedItem == -1) {
+            showAlert("Info", "No Question selected!");
+            return;
+        }
 
         Question q = selectedTopic.getAllQuestions().get(selectedItem);
         selectedTopic.deleteQuestion(q);
@@ -129,6 +150,15 @@ public class EditorController {
 
     }
 
+    private void showAlert(String title, String text) {
+        Alert alert = new Alert(Alert.AlertType.INFORMATION);
+        alert.setTitle(title);
 
+        // Header Text: null
+        alert.setHeaderText(null);
+        alert.setContentText(text);
+
+        alert.showAndWait();
+    }
 
 }
